% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains}
\alias{domains}
\title{Amazon domains other than .com}
\format{A list with countries and their urls and domains.
\describe{
  \item{url}{url to national Amazon}
  \item{domain}{domain to put into other functions}
  ...
}}
\usage{
domains
}
\description{
A data set listing the other possible Amazon domains
}
\keyword{datasets}

