% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reviews.R
\name{get_reviews}
\alias{get_reviews}
\title{Gets reviews from Amazon.com}
\usage{
get_reviews(productCode)
}
\arguments{
\item{productCode:}{The Amazon product code.}
}
\value{
The function returns a data frame with reviews.
}
\description{
The function gets reviews from Amazon.com for a particular product. It fetches the rating, the date of the review, the format if it is book, if the purchase was verified or not, the headline and the text of the review.
}
\examples{
# Get reviews for 'The Art of R Programming: A Tour of Statistical Software Design'

reviews <- Rmazon::get_reviews("1593273843")

}

